Move `tonic::verification` module to `common::strict_mode`
