+++ src/common/strict_mode.rs
